simplify lambda render tests
